// code/frontend/apps/game-index/src/gen/catalogApi/types/GameReleaseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { CompanyInfoDto } from "./CompanyInfoDto.ts";

export type GameReleaseDto = {
    /**
     * @type null,string
     */
    platformName?: string | null;
    /**
     * @type null,string
     */
    platformSlug?: string | null;
    /**
     * @type null,integer, int64
     */
    releaseDate?: number | null;
    /**
     * @type null,string
     */
    region?: string | null;
    /**
     * @type null,string
     */
    human?: string | null;
    /**
     * @type null,array
     */
    developers?: CompanyInfoDto[] | null;
    /**
     * @type null,array
     */
    publishers?: CompanyInfoDto[] | null;
};